pyproject.toml
[tool.poetry]
name = "advanced-rag"
version = "0.1.0"
description = ""
authors = ["Eyal Bar-Nir <eyal@chronom.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
beautifulsoup4 = "^4.12.3"
langchain = "^0.3.2"
langgraph = "^0.2.34"
langchainhub = "^0.1.21"
langchain-community = "^0.3.1"
tavily-python = "^0.5.0"
langchain-chroma = "^0.1.4"
python-dotenv = "^1.0.1"
black = "^24.8.0"
isort = "^5.13.2"
pytest = "^8.3.3"
langchain-openai = "^0.2.2"
pytest-dotenv = "^0.5.2"
langgraph-cli = "^0.1.52"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"